refactor: Update Tauri dependencies and plugins versions
