+++ package.json
@@ -21,6 +21,7 @@
     "@mantine/notifications": "^6.0.21",
     "@tabler/icons-react": "^2.47.0",
     "@tauri-apps/api": "^1.5.6",
+    "@tauri-apps/plugin-updater": "^2.0.0-beta.5",
     "ag-grid-react": "^31.3.2",
     "apexcharts": "^3.49.1",
     "classnames": "^2.5.1",
@@ -45,7 +46,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.4.0",
-    "@tauri-apps/cli": "^1.5.14",
+    "@tauri-apps/cli": "next",
     "@types/lodash.sortby": "^4.7.9",
     "@types/node": "^20.14.0",
     "@types/react": "^18.3.3",
