+++ src-tauri/Cargo.toml
@@ -17,7 +17,7 @@ members = [
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.5.2", features = [] }
+tauri-build = { version = "2.0.0-beta", features = [] }
 
 [dependencies]
 systemicons = { path = "./ws-systemicons" }
@@ -25,7 +25,7 @@ chrono = {version = "0.4.23", features= ["serde"] }
 # rusqlite = { version = "0.28.0", features = ["bundled", "chrono"] }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.6.7", features = ["api-all", "system-tray", "updater"] }
+tauri = { version = "2.0.0-beta", features = ["tray-icon"] }
 sysinfo = "0.30.12"
 tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
 tauri-plugin-store = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
@@ -41,6 +41,7 @@ fs_extra = "1.3.0"
 base64 = "0.22.1"
 surrealdb = { version = "1.5.2", features = ["kv-speedb"] }
 anyhow = "1.0.86"
+tauri-plugin-updater = "2.0.0-beta.7"
 
 
 [target.'cfg(target_os = "macos")'.dependencies]
@@ -61,6 +62,9 @@ features = ["winerror"]
 # surrealdb = { version = "1.5.0" }
 
 
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib", "lib"]
 
 [features]
 # by default Tauri runs in production mode
