+++ src/lib/helpers.ts
@@ -1,4 +1,4 @@
 import { Command } from '@/lib/types';
-import { invoke as invokeTauri, InvokeArgs } from '@tauri-apps/api/tauri';
+import { invoke as invokeTauri, InvokeArgs } from 'core';
 
 export const invoke = <T extends InvokeArgs, V = any>(cmd: Command, args?: T) => invokeTauri(cmd, args) as Promise<V>;
